Tables Relationships are Working Fine BUT Problem with Data input
All the tables relationships are set. However, when I test the application using the sign up form, no data is being stored in the tables. I tried to create a new database and store it there, but still the same error. I did make few changes, but I am still testing the sign up form with the database.