Login and Sign-Up DOES NOT WORK
Login and Sign-UP views are fine. Frontend looks good but Backend is giving me trouble. For some reason, when I try to test the application by entering data in the sign up page. The data is not being stored in the database. I cannot connect profile with login until sign up works and stores data for login